Removed private update menu from dotfiles

+++ .config/waybar/launch.sh
@@ -3,12 +3,13 @@
 DIR="${0%/*}"
 STATE_FILE="$DIR/.current"
 
-config=$(cat "$STATE_FILE" 2>/dev/null)
-
-pkill waybar
-
-if [[ -z "$config" || "$config" == "default" ]]; then
-    waybar &
+if pgrep -x waybar > /dev/null; then
+    pkill waybar
 else
-    waybar -c "$DIR/config-${config}.jsonc" -s "$DIR/style-${config}.css" &
+    config=$(cat "$STATE_FILE" 2>/dev/null)
+    if [[ -z "$config" || "$config" == "default" ]]; then
+        waybar &
+    else
+        waybar -c "$DIR/config-${config}.jsonc" -s "$DIR/style-${config}.css" &
+    fi
 fi
